schema validation moved to schemaUtils.js
