+++ utils/schemaUtils.js
@@ -1,3 +1,6 @@
+      // Initializing the AJV validator
+      const ajv = new Ajv();
+
       // function to extract from the datatype the location of the schema
       export function getSchemaLocation(jsonified, template = false) {
         if (!jsonified.datatype) {
@@ -98,4 +101,37 @@
         });
 
         return merged;
+      }
+
+
+      
+      export async function validateData(jsonified) {
+        // STEP 3: Check if the jsonified object is a valid JSON file against the schema
+        let [schema_type, schema_template] = await getSchemaType(jsonified);
+
+        if (Object.keys(schema_type).length === 0) {
+          errorDiv.innerText +=
+            "Schema check: There was no Schema specific to this DataType, or the schema was not compatible. The default scatter plot schema was used.\n";
+          schema_type = schema;
+        }
+
+        // validate the json
+        const validate = ajv.compile(schema_type);
+        const valid = validate(jsonified);
+
+        if (!valid) {
+          // Console log an error if the data is not valid against the schema
+          console.log("validate errors: ", JSON.stringify(validate.errors));
+          // Display an error message if the data is not valid against the schema
+          errorDiv.innerText += `Json file does not match the schema. ${JSON.stringify(validate.errors)}\n`;
+          errorDiv.innerText += `Json file does not match the schema. ${JSON.stringify(validate.errors)}\n`;
+          return null;
+        }
+
+        let _jsonified = jsonified;
+        if (Object.keys(schema_template).length !== 0) {
+          _jsonified = mergeFigDictWithTemplate(jsonified, schema_template);
+        }
+
+        return _jsonified;
       }
